+++ opal/core/search/tests/test_extract.py
@@ -217,14 +217,26 @@ class TestBasicCsvRenderer(PatientEpisodeTestCase):
 
     def test_get_field_names_to_render(self):
         with patch.object(Colour, "_get_fieldnames_to_extract") as field_names:
-            field_names.return_value = ["name", "consistency_token"]
+            field_names.return_value = ["id", "name", "consistency_token"]
             renderer = extract.CsvRenderer(
-                Colour,  Colour.objects.all(), self.user
+                Colour, Colour.objects.all(), self.user
             )
             self.assertEqual(
                 renderer.fields,
-                ["name"]
+                ["id", "name"]
             )
+
+    def test_get_field_names_to_render_order(self):
+        with patch.object(Colour, "_get_fieldnames_to_extract") as field_names:
+            field_names.return_value = ["name", "id", "patient_id", "consistency_token", "episode_id"]
+            renderer = extract.CsvRenderer(
+                Colour, Colour.objects.all(), self.user
+            )
+            self.assertEqual(
+                renderer.fields,
+                ["id", "patient_id", "episode_id", "name"]
+            )
+
 
     def test_fields_uses_fields_arg(self):
             renderer = extract.CsvRenderer(
@@ -440,20 +452,9 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         self.pid_str = str(self.patient.id)
         self.eid_str = str(self.episode.id)
 
-    def test_get_header(self, field_names_to_extract):
-        field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
-        ]
-        renderer = extract.PatientSubrecordCsvRenderer(
-            PatientColour,
-            models.Episode.objects.all(),
-            self.user
-        )
-        self.assertEqual(["Episode", "Patient", "Name"], renderer.get_headers())
-
     def test_get_row(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token"
         ]
         renderer = extract.PatientSubrecordCsvRenderer(
             PatientColour,
@@ -462,12 +463,17 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         )
         rendered = renderer.get_row(self.patient_colour, self.episode.id)
         self.assertEqual(
-            [str(self.episode.id), str(self.patient.id), "blue"], rendered
+            [
+                str(self.patient_colour.id),
+                self.pid_str,
+                self.eid_str,
+                "blue"
+            ], rendered
         )
 
     def test_get_rows(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token",
         ]
         renderer = extract.PatientSubrecordCsvRenderer(
             PatientColour,
@@ -477,7 +483,7 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         rendered = list(
             renderer.get_rows()
         )
-        expected = [[self.eid_str, self.pid_str, "blue"]]
+        expected = [[str(self.patient_colour.id), self.pid_str, self.eid_str, "blue"]]
         self.assertEqual(expected, rendered)
 
     def test_get_rows_same_patient(self, field_names_to_extract):
@@ -485,7 +491,7 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
         first_episode = self.patient.episode_set.first()
         last_episode = self.patient.episode_set.last()
         field_names_to_extract.return_value = [
-            "patient_id", "name", "consistency_token", "id"
+            "id", "patient_id", "episode_id", "name", "consistency_token"
         ]
 
         renderer = extract.PatientSubrecordCsvRenderer(
@@ -497,8 +503,18 @@ class TestPatientSubrecordCsvRenderer(PatientEpisodeTestCase):
             renderer.get_rows()
         )
         self.assertEqual([
-            [str(first_episode.id), str(self.patient.id), "blue"],
-            [str(last_episode.id), str(self.patient.id), "blue"]
+            [
+                str(self.patient_colour.id),
+                str(self.patient.id),
+                str(first_episode.id),
+                "blue"
+            ],
+            [
+                str(self.patient_colour.id),
+                str(self.patient.id),
+                str(last_episode.id),
+                "blue"
+            ],
         ], rendered)
 
 
@@ -512,26 +528,15 @@ class TestEpisodeSubrecordCsvRenderer(PatientEpisodeTestCase):
         self.eid_str = str(self.episode.id)
         self.pid_str = str(self.patient.id)
 
-    def test_get_header(self, field_names_to_extract):
-        field_names_to_extract.return_value = [
-            "episode_id", "name", "consistency_token", "id"
-        ]
-        renderer = extract.EpisodeSubrecordCsvRenderer(
-            Colour,
-            models.Episode.objects.all(),
-            self.user
-        )
-        self.assertEqual(["Patient", "Episode", "Name"], renderer.get_headers())
-
     def test_get_row(self, field_names_to_extract):
         field_names_to_extract.return_value = [
-            "episode_id", "name", "consistency_token", "id"
+             "id", "episode_id", "name", "consistency_token"
         ]
         renderer = extract.EpisodeSubrecordCsvRenderer(
             Colour, models.Episode.objects.all(), self.user
         )
         rendered = renderer.get_row(self.colour)
-        expected = [self.pid_str, self.eid_str, "blue"]
+        expected = [str(self.colour.id), self.pid_str, self.eid_str, "blue"]
         self.assertEqual(expected, rendered)