+++ changelog.md
@@ -2,10 +2,14 @@
 
 #### Application.modify_extract
 
-A list of functions attached to the application object that 
-enable the application to add, remove or alter files in an 
+A list of functions attached to the application object that
+enable the application to add, remove or alter files in an
 advanced search extract.
 
+
+#### Include ID in the extract
+
+The model ID will no longer be excluded from its csv extract. If present, id will appear first, then patient id, then episode id.
 
 ### 0.21.0 (Major Release)