+++ changelog.md
@@ -1,9 +1,12 @@
 ### 0.22.0 (Major Release)
 
+#### Advanced search changes
+Date and datetime fields now support `Equals` along with `Before` and `After`
+
 #### Application.modify_extract
 
-A list of functions attached to the application object that 
-enable the application to add, remove or alter files in an 
+A list of functions attached to the application object that
+enable the application to add, remove or alter files in an
 advanced search extract.