+++ changelog.md
@@ -10,6 +10,10 @@ enable the application to add, remove or alter files in an
 advanced search extract.
 
 
+#### Include ID in the extract
+
+The model ID will no longer be excluded from its csv extract. If present, id will appear first, then patient id, then episode id.
+
 ### 0.21.0 (Major Release)
 
 #### Celery upgrade and Django Celery library change.